Simple performance tests for site rating view
