+++ SpeedTests/SiteRatingViewPerfTests.swift
@@ -1,0 +1,48 @@
+//  SiteRatingViewPerfTests.swift
+//  DuckDuckGo
+//
+//  Copyright © 2019 DuckDuckGo. All rights reserved.
+//
+//  Licensed under the Apache License, Version 2.0 (the "License");
+//  you may not use this file except in compliance with the License.
+//  You may obtain a copy of the License at
+//
+//  http://www.apache.org/licenses/LICENSE-2.0
+//
+//  Unless required by applicable law or agreed to in writing, software
+//  distributed under the License is distributed on an "AS IS" BASIS,
+//  WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+//  See the License for the specific language governing permissions and
+//  limitations under the License.
+//
+
+import XCTest
+
+@testable import DuckDuckGo
+@testable import Core
+
+class SiteRatingViewPerfTests: XCTestCase {
+    
+    private var mainController: MainViewController!
+
+    override func setUp() {
+        loadBlockingLists()
+        TabsModel.clear()
+        mainController = loadStoryboard()
+    }
+    
+    override func tearDown() {
+        TabsModel.clear()
+    }
+
+    func testSiteRatingViewRefreshPerformance() {
+        
+        mainController.loadUrl(URL(string: "https:/duckduckgo.com")!)
+        waitForPageLoad(in: mainController)
+        
+        self.measure {
+            self.mainController.omniBar.siteRatingView.refresh()
+        }
+    }
+
+}
